Updated Build documentation in Readme
When I applied a fix for Toolchains, I noticed that the pom's plugins were configured incorrectly.
We were using <build><pluginManagement><plugins>...
The <pluginManagement> element is only used in a Parent POM. since our pom structure is the main pom, we just needed <build><plugins>...
This enabled all of the mapped tasks such as assembly:single to work as intended.
mvn compile will now build the jar-with-dependencies.