Move Toolchains Plugin to profile
Toolchains plugin is now configured as a profile which is active unliss an environment variable "TRAVIS" is defined.
This allows toolchains under normal builds, but skips toolchains when running on travis-ci.org (free CI server)
We may want to consider making this more portable (less coupled to travis) by setting a different trigger.